pyproject.toml
[project]
name = "ironsitehackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.83.0",
    "controlnet-aux>=0.0.10",
    "diffusers>=0.36.0",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
]
